improving exceptions thrown during transformation
[minor][change]
Due to exceptions being raised in kind of "random" positions, improving and streamlining them was needed. Eg. if a transformer failed due to some syntax error that was not handled correctly, it may not be findable without heavy testing. This has changed due to inclusion of an additional try-catch to find those unhandled transformer errors and report them accordingly.