fixing transformation scope stalling in certain templates
[minor][fix]
This problem got reported to me via mail after releasing the previous version. The background here has to do with variable scopes not living through the whole "time" they actually have to but rather live in a quirky way, to enable some fancy things.

The solution was to make the variable scopes less quirky.