add default fallback namespace
[minor][change]
Due to the fact that the xmlns was always mandatory for the system to actually find something, a default behavior was added.
This now allows to omit the xmlns definition, making things easier to place (eg. the xmlns had to be placed on the very first element and may never have been placed on body/header/footer because they depend on the template parameter)