Refactor TextControl and introduce TextBaseControl
The TextControl class has been refactored and a new class TextBaseControl has been introduced. Most of the properties and methods related to text styling and rendering have been shifted to the TextBaseControl class. The aim for this refactoring is to have a base class for all controls intended to display simple text.