Added README.md to the NuGet package information
The PackageReadmeFile tag in the PdfTemplate.csproj file has been updated to include 'README.md'. Also, a path to the README.md file has been added to the ItemGroup for NuGet packaging, ensuring it's included when the package is built.

This might not work tho ... so lets hope that the nuget.org page starts to shine in a bit :D