Fixed style order
A new test for applying the lowest level style last has been added in XmlTemplateReaderTests.cs. To ensure that this test passes and the style implementation is correct, the code for processing styles in XmlTemplateReader.cs has been updated to reverse the order of styles. In addition, the version number has been bumped to 4.0.1 in the main GitHub workflow.