Enable short alternation using `@alternate on value`
This update modifies the condition in the AlternateTransformer class to avoid SequenceEqual check when the values array has no elements. Additionally, the value assignation from 'values' is changed to 'alternateValue.Values' for consistency with the rest of the code. The change, effectively, enables `@alternates on value` to be used instead of having to define the value each time. Additionally, a mistake in the README.md was fixed with the table headers.