Fix background page height in PDF generator.
Removed adjustment to background page height to use full original page height.
No version bump is needed because the commit was not yet sent to CI/CD. Due to a workflow issue of mine (forgot i updated the project on github ...) and to keep the history in tact, this was added after the worklow version bump.