Add support for foreground controls in PDF generation.
Introduced foreground controls to render elements on top of other content on each page. Updated the Generator and Template classes accordingly, and added unit tests to validate the new functionality.
No version bump is needed because the commit was not yet sent to CI/CD. Due to a workflow issue of mine (forgot i updated the project on github ...) and to keep the history in tact, this was added after the worklow version bump.