Fix bitmaps being created transparent and update ReadMe
This update fixes the ReadMe TOC (was not updated after the last change) and, additionally, fixes a debug artifact i introduced by accident (canvas.Clear(Transparent)), which resulted in working PDFs but for BitMaps generated to have a transparent background. The Fix should have been up already, but apparently I forgot to push this. Sorry for the inconvenience.